vpn_bot_copy/requirements.txt
python-telegram-bot==20.7
requests==2.31.0
python-dotenv==1.0.0
qrcode==7.4.2
Pillow==10.0.1